v23669 (06/25/01):
- (bug fix) Vehicles and deployables now properly explode if they are destroyed while someone is repairing them.

- (bug fix) Sniper laser shots no longer create water splash effects if hitting ground near the water.

- (bug fix) Immersion iForce force feedback mouse is now working properly again.

- (bug fix) The "flag jumping" bug is now fixed. When a flag lands after being dropped, it will stay put when it slides to rest.

- (bug fix) Fixed a situation where closing tribe or player tags on the Browser out of order would cause a disconnect with the database server.

- (bug fix) Players can no longer fire, place mines, place grenades, or place beacons when inside a force field.

- (bug fix) Fixed bug where modifier keys (specifically SHIFT) bound to actions in the game would still cause those actions when typing in a text edit control in-game (such as the chat entry)

- (bug fix) Fixed a bug that could cause a player to drop to desktop when attempting to join a game which was in the process of cycling missions.

- (bug fix) Fixed a Radeon video card issue which could occur if the desktop color bit-depth was different than the color bit-depth that the player was using in the game.

- (bug fix) You won't try to fade into a vehicle that was destroyed after you purchased it, but before you had actually tported to the seat.

- (bug fix) Minor change in the MPB explosion so that the turret part of the MPB doesn't seem to hover in place for a split-second during the explosion.

- (bug fix) Bomber bombs now tumble properly and won't seem to disappear when falling.

- (bug fix) Fixed a rare problem that could cause a client crash while the server is resetting.

- (bug fix) Fixed a problem with the ELF gun effect that was causing hangs.

- (bug fix) Telnet can now be used to set passwords for PURE servers so that they can be used for match games. (command line option...see the post in T2FAQs called "How do I TELNET INTO A PURE SERVER?" for more information on how to use this ability.)

- (bug fix) The "cloning" issue (where players could clone themselves by dying, going to the CC with the CC camera showing themselves, spawn, and flicker back and forth to the CC) is now fixed and no longer occurs.

- (bug fix) Another "cloning" issue which occurred when the Tourney Admin would switch teams for players is fixed and no longer occurs.

- (bug fix) Fixed a bug where, when a client joins a server where the client does not have the map being run on the server, the client hangs while loading. This now elegantly exits instead of hanging.

- (bug fix) Fixed an issue where the last few characters of the Server Info dialog would be cut off.

- (bug fix) Fixed a situation where a blank error box could occur if CD key not entered properly when creating an account.

- (bug fix) Sensor rings will no longer show up on the Command Circuit if the generators are not powered.

- (bug fix) There was a rare bug where, if a player was standing in a force field's position when that force field went from a depowered to powered state (in otherwords, if the gens were repaired while he stood in the FF position), then the player would be stuck forever. If this case occurs, that player will now be destroyed.

- (improvement) The "redjack" icon has been removed and new network throughput graphs have been implemented to better help players troubleshoot their net settings. Additionally, a more accurate and complete set of network presets is available. (See details below under "NETWORK SETTINGS" for more information.)

- (improvement) The pure server concept is now implemented. Pure servers only allow regulation scripts and maps to be run on the server (no restrictions yet on the client), thus ensuring that anyone that joins a "base" server is playing the game as it was designed by Dynamix. MODs are still easily joined, but players can be assured that a "Base" game is really a "Base" game now. (Any game with server or rules mods that is not actually named as a new MOD will be described as "variant" instead of "base" on the master server list.) NOTE: See "PURE SERVER" below for more information on this.

- (improvement) Old Password is now required in order to enter a New Password when editing your account.

- (improvement) Password handling is different now in order to make it more difficult for people to casually find a password on a hard drive.

- (improvement) Made more room for player names to display on the Server Info box.

- (improvement) Bomber and Tank now have separate energy capacitors for their turret weapons. This energy pool is completely separate from the energy pool that the thrusters and force shields use. (Gunner energy is displayed as a second bar below the regular vehicle energy and is orangish in color.)

- (improvement) Vertical thrusters on air vehicles are now more efficient to enable better takeoffs from ground level.

- (improvement) Changed team damage OFF to include friendly turret fire and vehicle fire. (In otherwords, if Team Damage is OFF, then turrets fire and vehicle fire will not affect friendly units.)

- (improvement) Splash damage no longer falls off so dramatically with distance. You will find that area effect (explosion) weapons now are more effective within their damage area.

- (improvement) While in Tournament mode, and while in observer mode at the beginning of a match (before teams have been selected), players will now be able to chat with one another.

- (improvement) Added observer points to Tombstone (it previously had none).

- (improvement) Grenades tossing is slightly improved. The grenades will throw farther with less time spent pressing the grenade key. (They still have the exact same minimum and maximum throwing distances, it's just easier to throw it out to max range now.)

- (community) Player histories are now accurate.

- (community) Preferences in the FORUMs should be fixed now so that they stay in existence. The sort is the only exception. That will reformat each time you enter the FORUM and you will need to select whichever sort your prefer at that time.

+++ base/scripts/GameGui.cs
@@ -18,7 +18,7 @@ function GameGui::onWake( %this )
 {
    Canvas.pushDialog( LaunchToolbarDlg );
 
-	if ( isDemo() || $PlayingOnline )
+	if ( isDemo() || isDemoServer() || $PlayingOnline )
    	GM_Frame.setTitle( "GAME" );
 	else
    	GM_Frame.setTitle( "LAN GAME" );
@@ -26,15 +26,23 @@ function GameGui::onWake( %this )
    // This is essentially an "isInitialized" flag...
    if ( GM_TabView.tabCount() == 0 )
    {
-      GM_TabView.addTab( 1, "JOIN" );
-		if ( !isDemo() )
+      if ( isDemo() )
+      {
+         GM_TabView.addTab( 1, "JOIN" );
+         %this.pane = "Join";
+      }
+      else if ( isDemoServer() )
+      {
+         GM_TabView.addTab( 2, "HOST" );
+         %this.pane = "Host";
+      }
+		else
 		{
+         GM_TabView.addTab( 1, "JOIN" );
       	GM_TabView.addTab( 2, "HOST" );
       	GM_TabView.addTab( 3, "WARRIOR SETUP", 1 );
       	queryMasterGameTypes();
 		}
-		else
-			%this.pane = "Join";
    }
 
    switch$ ( %this.pane )
@@ -130,7 +138,10 @@ function GM_JoinPane::onActivate( %this )
       BrowserMap.delete();
    }
    new ActionMap( BrowserMap );
-   BrowserMap.bindCmd( keyboard, insert, "GMJ_Browser.insertIPAddress();", "" );
+   if ( !isDemo() )
+      BrowserMap.bindCmd( keyboard, insert, "GMJ_Browser.insertIPAddress();", "" );
+   BrowserMap.bindCmd( keyboard, "ctrl f", "Canvas.pushDialog( FindServerDlg );", "" );
+   BrowserMap.bindCmd( keyboard, F3, "GMJ_Browser.findNextServer();", "" );
    BrowserMap.push();
 
    GM_VersionText.setVisible( !isDemo() );
@@ -142,8 +153,11 @@ function GM_JoinPane::onActivate( %this )
 //------------------------------------------------------------------------------
 function GM_JoinPane::onDeactivate( %this )
 {
-   BrowserMap.pop();
-   BrowserMap.delete();
+   if ( isObject( BrowserMap ) )
+   {
+      BrowserMap.pop();
+      BrowserMap.delete();
+   }
    
    GM_VersionText.setVisible( false );
 
@@ -172,9 +186,12 @@ $BrowserColumnCount++;
 $BrowserColumnName[5] = "Mission Name";
 $BrowserColumnRange[5] = "25 300";
 $BrowserColumnCount++;
-$BrowserColumnName[6] = "Rules Set";
-$BrowserColumnRange[6] = "25 300";
-$BrowserColumnCount++;
+if ( !isDemo() )
+{
+   $BrowserColumnName[6] = "Rules Set";
+   $BrowserColumnRange[6] = "25 300";
+   $BrowserColumnCount++;
+}
 $BrowserColumnName[7] = "# Players (Bots)";
 $BrowserColumnRange[7] = "25 150";
 $BrowserColumnCount++;
@@ -488,10 +505,37 @@ function EnterIPDlg::onDone( %this )
       %address = "IP:" @ %address;
    if ( strpos( %address, ":", 3 ) == -1 )
       %address = %address @ ":28000";
-      
-   error( ">> address = \"" @ %address @ "\" <<" );
+   
+   echo( "Starting ping to server " @ %address @ "..." );   
    pushServerAddress( %address );
-   GMJ_Browser.selectRowByAddress( %address );
+   GMJ_Browser.selectRowByAddress( %address, true );
+}
+
+//------------------------------------------------------------------------------
+function FindServerDlg::onWake( %this )
+{
+   FS_SearchPattern.validate();
+   FS_SearchPattern.selectAll();
+}
+
+//------------------------------------------------------------------------------
+function FindServerDlg::onGo( %this )
+{
+   %pattern = FS_SearchPattern.getValue();
+   if ( %pattern !$= "" )
+   {
+      Canvas.popDialog( FindServerDlg );
+      if ( !GMJ_Browser.findServer( %pattern ) )
+         MessageBoxOK( "NOT FOUND", "No servers with \"" @ %pattern @ "\" in their name were found." );
+   }
+   else
+      alxPlay( InputDeniedSound, 0, 0, 0 );
+}
+
+//------------------------------------------------------------------------------
+function FS_SearchPattern::validate( %this )
+{
+   FS_GoBtn.setActive( %this.getValue() !$= "" );
 }
 
 //------------------------------------------------------------------------------
@@ -558,7 +602,7 @@ function ServerInfoDlg::update( %this )
       %temp = getRecord( %info, 5 );
       if ( %temp !$= "" )
          %infoText = %infoText NL "<spush>" @ %this.headerStyle @ "MAP NAME:<spop><lmargin:70>" TAB %temp @ "<lmargin:0>";
-      %temp = getRecord( %info, 6 );
+      %temp = getRecords( %info, 6, 10 );
       if ( %temp !$= "" )
          %infoText = %infoText NL "<spush>" @ %this.headerStyle @ "SERVER INFO:<spop><lmargin:10>" TAB %temp @ "<lmargin:0>";
 
@@ -607,22 +651,22 @@ function SI_ContentWindow::fill( %this, %content )
    %record++;
    if ( %teamCount > 1 )
    {
-      %string = "<tab:110,220><spush>" @ ServerInfoDlg.headerStyle @ "TEAMS" TAB "SCORE<spop><tab:5,115,225>";
+      %string = "<spush>" @ ServerInfoDlg.headerStyle @ "TEAMS<lmargin%:50>SCORE<spop>";
       for ( %i = 0; %i < %teamCount; %i++ )
       {
          %teamEntry = getRecord( %content, %record );
-         %string = %string NL "<clip:110>\t" @ getField( %teamEntry, 0 ) @ "</clip>" TAB getField( %teamEntry, 1 );
+         %string = %string NL "<lmargin:0><clip%:50>" SPC getField( %teamEntry, 0 ) @ "</clip><lmargin%:50>" SPC getField( %teamEntry, 1 );
          %record++;
       }
 
       %playerCount = getRecord( %content, %record );
       %record++;
-      %string = %string NL "\n<tab:110,220><spush>" @ ServerInfoDlg.headerStyle @ "PLAYERS" TAB "TEAM" TAB "SCORE<spop><tab:5,115,225>";
+      %string = %string NL "\n<lmargin:0><spush>" @ ServerInfoDlg.headerStyle @ "PLAYERS<lmargin%:40>TEAM<lmargin%:75>SCORE<spop>";
       for ( %i = 0; %i < %playerCount; %i++ )
       {
          %playerEntry = getRecord( %content, %record ); 
-         %string = %string NL "<clip:110>\t" @ getField( %playerEntry, 0 ) @ "</clip>" 
-               TAB "<clip:105>" @ getField( %playerEntry, 1 ) @ "</clip>" TAB getField( %playerEntry, 2 );
+         %string = %string NL "<lmargin:0><clip%:40>" SPC getField( %playerEntry, 0 ) @ "</clip><lmargin%:40><clip%:35>" 
+               SPC getField( %playerEntry, 1 ) @ "</clip><lmargin%:75><clip%:25>" SPC getField( %playerEntry, 2 ) @ "</clip>";
          %record++;
       }
    }
@@ -631,11 +675,11 @@ function SI_ContentWindow::fill( %this, %content )
       %record++;
       %playerCount = getRecord( %content, %record );
       %record++;
-      %string = "<tab:150><spush>" @ ServerInfoDlg.headerStyle @ "PLAYERS" TAB "SCORE<spop><tab:5,155>";
+      %string = "<spush>" @ ServerInfoDlg.headerStyle @ "PLAYERS<lmargin%:60>SCORE<spop>";
       for ( %i = 0; %i < %playerCount; %i++ )
       {
          %playerEntry = getRecord( %content, %record ); 
-         %string = %string NL "<clip:150>\t" @ getField( %playerEntry, 0 ) @ "</clip>" TAB getField( %playerEntry, 2 );
+         %string = %string NL "<lmargin:0><clip%:60>" SPC getField( %playerEntry, 0 ) @ "</clip><lmargin%:60>" SPC getField( %playerEntry, 2 );
          %record++;
       }
    }
@@ -689,6 +733,7 @@ function JoinGame(%address)
 	%playerSkin = getField( %playerPref, 2 );
 	%playerVoice = getField( %playerPref, 3 );
 	%playerVoicePitch = getField( %playerPref, 4 );
+   LoadingGui.gotLoadInfo = "";
    connect( %address, $JoinGamePassword, %playerName, %playerRaceGender, %playerSkin, %playerVoice, %playerVoicePitch );
 }
 
@@ -758,6 +803,7 @@ function buildMissionTypePopup( %popup )
    %popup.clear();
    for( %type = 0; %type < $HostTypeCount; %type++ )
       %popup.add( $HostTypeDisplayName[%type], %type );
+   %popup.sort( true );   
 }
 
 //------------------------------------------------------------------------------
@@ -994,8 +1040,8 @@ function StartHostedGame()
    if ( $Host::Dedicated )
    {
       MessageBoxYesNo( "WARNING", 
-            "You are about to launch a dedicated server and leave Tribes 2.  Do you want to continue?", 
-            "tryToLaunchDedicatedServer();" );
+            "You are about to launch a dedicated server and quit Tribes 2.  Do you want to continue?", 
+            "tryToLaunchDedicatedServer(" @ $Host::PureServer @ ");" );
       return;
    }
 
@@ -1023,10 +1069,10 @@ function StartHostedGame()
 }
 
 //------------------------------------------------------------------------------
-function tryToLaunchDedicatedServer()
+function tryToLaunchDedicatedServer( %pure )
 {
    %numBots = $Host::BotsEnabled ? $Host::BotCount : 0;
-   if ( launchDedicatedServer( $Host::MissionType, $Host::Map, %numBots ) )
+   if ( launchDedicatedServer( $Host::MissionType, $Host::Map, %numBots, %pure ) )
       quit();
    else
    {
@@ -1162,6 +1208,7 @@ function AdvancedHostDlg::onWake( %this )
    // Set all of the controls to the current pref states:
    AH_HostPort.setText( $Host::Port );
    AH_DedicatedTgl.setValue( $Host::Dedicated );
+   AH_DedicatedTgl.onAction();
    AH_TournamentTgl.setValue( $Host::TournamentMode );
    AH_AdminVoteTgl.setValue( $Host::allowAdminPlayerVotes );
    AH_AllowSmurfTgl.setValue( !$Host::NoSmurfs );
@@ -1180,6 +1227,8 @@ function AdvancedHostDlg::accept( %this )
    // Apply all of the changes:
    $Host::Port = AH_HostPort.getValue();
    $Host::Dedicated = AH_DedicatedTgl.getValue();
+   if ( $Host::Dedicated )
+      $Host::PureServer = AH_PureServerTgl.getValue();
    $Host::TournamentMode = AH_TournamentTgl.getValue();
    $Host::allowAdminPlayerVotes = AH_AdminVoteTgl.getValue();
    $Host::NoSmurfs = !AH_AllowSmurfTgl.getValue();
@@ -1195,6 +1244,21 @@ function AdvancedHostDlg::accept( %this )
    export( "$Host::*", "prefs/ServerPrefs.cs", false );
 
    Canvas.popDialog( AdvancedHostDlg );
+}
+
+//------------------------------------------------------------------------------
+function AH_DedicatedTgl::onAction( %this )
+{
+   if ( %this.getValue() )
+   {
+      AH_PureServerTgl.setValue( $Host::PureServer );
+      AH_PureServerTgl.setActive( true );
+   }
+   else
+   {
+      AH_PureServerTgl.setValue( false );
+      AH_PureServerTgl.setActive( false );
+   }
 }
 
 //------------------------------------------------------------------------------
@@ -1522,6 +1586,8 @@ function GMW_SkinPopup::fillList( %this, %raceGender )
          %count++;
       }
    }
+   
+   %this.sort( true );
 }
 
 //------------------------------------------------------------------------------
